:card_file_box: Add Ability table and code related to it
Note: I'll probably will have to delete the PokemonResponseToModelMapper and just use Room relations to fetch the complete metadata. Anyways, this MR creates the logic that adds the Ability table and for each ability in the /pokemon API it tries to figure it out if the ability already exists or not. If it does, then it doesnt do anything. If it doesnt, then it adds to the database. I also had to mock the list of abilities in the PokemonModel because it'll be replaced by Room's relation model.